+++ built/base/AbstractNode.d.ts
@@ -1,5 +1,5 @@
 declare let _jtreeLatestTime: number;
 declare let _jtreeMinTimeIncrement: number;
 declare abstract class AbstractNode {
-    protected _getNow(): number;
+    protected _getProcessTimeInMilliseconds(): number;
 }