+++ built/base/AbstractNode.node.js
@@ -1,8 +1,9 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 class AbstractNode {
-    _getNow() {
-        return parseFloat(process.hrtime().join(""));
+    _getProcessTimeInMilliseconds() {
+        const hrtime = process.hrtime();
+        return (hrtime[0] * 1e9 + hrtime[1]) / 1e6;
     }
 }
 exports.default = AbstractNode;