+++ built/base/TreeNode.js
@@ -1222,7 +1222,7 @@ class TreeNode extends ImmutableNode {
         if (cmTime)
             mTimes.push(cmTime);
         const newestTime = Math.max.apply(null, mTimes);
-        return this._setCMTime(newestTime || this._getNow())._getCMTime();
+        return this._setCMTime(newestTime || this._getProcessTimeInMilliseconds())._getCMTime();
     }
     _getCMTime() {
         return this._cmtime;
@@ -1346,7 +1346,7 @@ class TreeNode extends ImmutableNode {
         return this._setChildren(children);
     }
     _updateMTime() {
-        this._mtime = this._getNow();
+        this._mtime = this._getProcessTimeInMilliseconds();
     }
     insertWord(index, word) {
         const wi = this.getZI();
@@ -1468,7 +1468,7 @@ class TreeNode extends ImmutableNode {
         this._clearIndex();
         // note: assumes indexesToDelete is in ascending order
         indexesToDelete.reverse().forEach(index => this._getChildrenArray().splice(index, 1));
-        return this._setCMTime(this._getNow());
+        return this._setCMTime(this._getProcessTimeInMilliseconds());
     }
     _deleteNode(node) {
         const index = this._indexOfNode(node);