+++ src/base/AbstractNode.node.ts
@@ -1,6 +1,7 @@
 abstract class AbstractNode {
-  protected _getNow() {
-    return parseFloat(process.hrtime().join(""))
+  protected _getProcessTimeInMilliseconds() {
+    const hrtime = process.hrtime()
+    return (hrtime[0] * 1e9 + hrtime[1]) / 1e6
   }
 }